# Tinkering with MPI launch invocations