@timmoon10 2019-10-21
[Bamboo passes.](https://lc.llnl.gov/bamboo/browse/LBANN-TIM225-2)